feat(workspace): Deduplicate deposit source types
