+++ crates/consensus/src/deposit/source.rs
@@ -4,7 +4,7 @@ use alloc::string::String;
 use alloy_primitives::{keccak256, B256};
 
 /// Source domain identifiers for deposit transactions.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Copy)]
 #[repr(u8)]
 pub enum DepositSourceDomainIdentifier {
     /// A user deposit source.
@@ -13,6 +13,8 @@ pub enum DepositSourceDomainIdentifier {
     L1Info = 1,
     /// An upgrade deposit source.
     Upgrade = 2,
+    /// Deposit context closing transaction.
+    DepositContext = 3,
 }
 
 /// Source domains for deposit transactions.
@@ -24,6 +26,8 @@ pub enum DepositSourceDomain {
     L1Info(L1InfoDepositSource),
     /// An upgrade deposit source.
     Upgrade(UpgradeDepositSource),
+    /// A deposit context closing source
+    DepositContext(DepositContextDepositSource),
 }
 
 impl DepositSourceDomain {
@@ -33,12 +37,42 @@ impl DepositSourceDomain {
             Self::User(ds) => ds.source_hash(),
             Self::L1Info(ds) => ds.source_hash(),
             Self::Upgrade(ds) => ds.source_hash(),
+            Self::DepositContext(ds) => ds.source_hash(),
         }
     }
 }
 
+/// A deposit transaction source.
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Copy)]
+pub struct UserDepositSource {
+    /// The L1 block hash.
+    pub l1_block_hash: B256,
+    /// The log index.
+    pub log_index: u64,
+}
+
+impl UserDepositSource {
+    /// Creates a new [UserDepositSource].
+    pub const fn new(l1_block_hash: B256, log_index: u64) -> Self {
+        Self { l1_block_hash, log_index }
+    }
+
+    /// Returns the source hash.
+    pub fn source_hash(&self) -> B256 {
+        let mut input = [0u8; 32 * 2];
+        input[..32].copy_from_slice(&self.l1_block_hash[..]);
+        input[32 * 2 - 8..].copy_from_slice(&self.log_index.to_be_bytes());
+        let deposit_id_hash = keccak256(input);
+        let mut domain_input = [0u8; 32 * 2];
+        let identifier_bytes: [u8; 8] = (DepositSourceDomainIdentifier::User as u64).to_be_bytes();
+        domain_input[32 - 8..32].copy_from_slice(&identifier_bytes);
+        domain_input[32..].copy_from_slice(&deposit_id_hash[..]);
+        keccak256(domain_input)
+    }
+}
+
 /// A L1 info deposit transaction source.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Copy)]
 pub struct L1InfoDepositSource {
     /// The L1 block hash.
     pub l1_block_hash: B256,
@@ -61,35 +95,6 @@ impl L1InfoDepositSource {
         let mut domain_input = [0u8; 32 * 2];
         let identifier_bytes: [u8; 8] =
             (DepositSourceDomainIdentifier::L1Info as u64).to_be_bytes();
-        domain_input[32 - 8..32].copy_from_slice(&identifier_bytes);
-        domain_input[32..].copy_from_slice(&deposit_id_hash[..]);
-        keccak256(domain_input)
-    }
-}
-
-/// A deposit transaction source.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct UserDepositSource {
-    /// The L1 block hash.
-    pub l1_block_hash: B256,
-    /// The log index.
-    pub log_index: u64,
-}
-
-impl UserDepositSource {
-    /// Creates a new [UserDepositSource].
-    pub const fn new(l1_block_hash: B256, log_index: u64) -> Self {
-        Self { l1_block_hash, log_index }
-    }
-
-    /// Returns the source hash.
-    pub fn source_hash(&self) -> B256 {
-        let mut input = [0u8; 32 * 2];
-        input[..32].copy_from_slice(&self.l1_block_hash[..]);
-        input[32 * 2 - 8..].copy_from_slice(&self.log_index.to_be_bytes());
-        let deposit_id_hash = keccak256(input);
-        let mut domain_input = [0u8; 32 * 2];
-        let identifier_bytes: [u8; 8] = (DepositSourceDomainIdentifier::User as u64).to_be_bytes();
         domain_input[32 - 8..32].copy_from_slice(&identifier_bytes);
         domain_input[32..].copy_from_slice(&deposit_id_hash[..]);
         keccak256(domain_input)
@@ -123,6 +128,36 @@ impl UpgradeDepositSource {
             (DepositSourceDomainIdentifier::Upgrade as u64).to_be_bytes();
         domain_input[32 - 8..32].copy_from_slice(&identifier_bytes);
         domain_input[32..].copy_from_slice(&intent_hash[..]);
+        keccak256(domain_input)
+    }
+}
+
+/// A deposit context transaction source.
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Copy)]
+pub struct DepositContextDepositSource {
+    /// The L1 block hash.
+    pub l1_block_hash: B256,
+    /// The sequence number.
+    pub seq_number: u64,
+}
+
+impl DepositContextDepositSource {
+    /// Creates a new [L1InfoDepositSource].
+    pub const fn new(l1_block_hash: B256, seq_number: u64) -> Self {
+        Self { l1_block_hash, seq_number }
+    }
+
+    /// Returns the source hash.
+    pub fn source_hash(&self) -> B256 {
+        let mut input = [0u8; 32 * 2];
+        input[..32].copy_from_slice(&self.l1_block_hash[..]);
+        input[32 * 2 - 8..].copy_from_slice(&self.seq_number.to_be_bytes());
+        let deposit_id_hash = keccak256(input);
+        let mut domain_input = [0u8; 32 * 2];
+        let identifier_bytes: [u8; 8] =
+            (DepositSourceDomainIdentifier::DepositContext as u64).to_be_bytes();
+        domain_input[32 - 8..32].copy_from_slice(&identifier_bytes);
+        domain_input[32..].copy_from_slice(&deposit_id_hash[..]);
         keccak256(domain_input)
     }
 }
