+++ crates/protocol/src/info/variant.rs
@@ -5,13 +5,10 @@ use alloc::{format, string::ToString};
 use alloy_consensus::Header;
 use alloy_eips::{eip7840::BlobParams, BlockNumHash};
 use alloy_primitives::{address, Address, Bytes, Sealable, Sealed, TxKind, B256, U256};
-use maili_consensus::TxDeposit;
+use maili_consensus::{DepositSourceDomain, L1InfoDepositSource, TxDeposit};
 use maili_genesis::{RollupConfig, SystemConfig};
 
-use crate::{
-    BlockInfoError, DecodeError, DepositSourceDomain, L1BlockInfoBedrock, L1BlockInfoEcotone,
-    L1InfoDepositSource,
-};
+use crate::{BlockInfoError, DecodeError, L1BlockInfoBedrock, L1BlockInfoEcotone};
 
 use super::L1BlockInfoInterop;
